Character Top/Bottom (READ DESCRIPTION)
A message for anyone using these images: This isn't the final character
design, but it's something you guys can easily work with for now! The
dimensions are 460x730 pixels and you can set the axis of rotation for
the top half to the coordinates (130, 430) to be able to have him look
somewhat natural when aiming his laser at any angle between 90 and -90
degrees. I know this isn't what you wanted Billy, but for a head/arm
only rotation to work, we would need 3 separate images as opposed to the
two you requested since the heads and the arms would operate on
different axes of rotation. If the programmers can work with 3 different
images and handle two separate axes of rotation then I will provide a
suitable set of images posthaste if requested. Let me know if you have
comments or questions!