Added comments to ShootingScript and TargetScript
##PLEASE READ##
The TargetScript is and should be attached to every enemy to define its
color.  This color is used to link it to the ShootingScript so that the
player's JKL keys are linked to the respective targets.
Please read the scripts, I added comments for that reason
##END PLEASE READ##

Added some comments so that you guys can work today (Friday) because I
won't be able to come in today~
Things to do:
Re-add enemyMovement (I didn't touch the enemyMovement script - besides
maybe reallocating the initcolor into the targetScript).
Map out the enemy locations onto the map.
If feeling ambitious, try to add another line renderer that is much
bolder for when the player shoots, but I'll try to do this over the
weekend since I learned some tricks from making the first lineRenderers