Starting implement music in Pirate Mountain
Rewrote the BgmPlayer script to allow for multiple different tracks in a
level, as well as made basic systems to allow flags in the level to
trigger different music. Once Mark's map is put in I'll start testing
all the systems (which are probably very broken right now)

I also put in a few lines to play the "player hurt" sfx in
EnemyPreferences, as well as added a few more audio files to Assets.