Enemy Shooting Nearly Complete
All BLUE caterpillars have a shooting script attached to them.  I'm
waiting to flesh out the rest of the code before adding the script to
all enemies... which is sort of an annoying job too.
Things left to do:
Bullets need to destroy themselves once they come in contact with a
Collider (while avoiding other collisions with other enemies and
possibly their bullets).